+++ src/app/HealthChecks/JsonResponseWriter.cs
@@ -1,0 +1,23 @@
+using Microsoft.AspNetCore.Http;
+using Microsoft.Extensions.Diagnostics.HealthChecks;
+using System.Text.Json;
+using System.Threading.Tasks;
+
+namespace Helium
+{
+    public partial class CosmosHealthCheck : IHealthCheck
+    {
+        /// <summary>
+        /// Write the health check results as json
+        /// </summary>
+        /// <param name="httpContext">HttpContext</param>
+        /// <param name="healthReport">HealthReport</param>
+        /// <returns></returns>
+        public static Task JsonResponseWriter(HttpContext httpContext, HealthReport healthReport)
+        {
+            // write the json
+            httpContext.Response.ContentType = "application/json";
+            return httpContext.Response.WriteAsync(JsonSerializer.Serialize(healthReport, jsonOptions));
+        }
+    }
+}